requirements.txt
fastapi==0.109.1
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.28.2
python-multipart==0.0.6
tonconnect==0.2.1  # Правильное название пакета!
psycopg2-binary==2.9.6  # Только для PostgreSQL